Add files via upload
The area shown next to the "ncocr"  files or in black and peach, are showing the non-coding open chromatin in the genome. The pink bar is the open chromatin regions. These regions were analyzed and located using using ATAC sequencing. Gene expression is shown in the dark blue, and green columns.